Modernize toolchain, dependencies, and lint to clean build
Upgrade the project to AGP 9.2.1, Gradle 9.5.1, Kotlin 2.4.0, and KSP
2.3.9. Migrate off transitional AGP/Kotlin flags and the standalone
kotlin-android plugin in favor of AGP 9 built-in Kotlin.

Bump AndroidX and Compose dependencies: compileSdk/targetSdk 37,
core-ktx 1.19.0, lifecycle 2.10.0, activity-compose 1.13.0,
compose-bom 2026.05.01, material3 1.5.0-alpha21, datastore 1.2.1,
work 2.11.2, jsoup 1.22.2, and androidx-junit 1.3.0. Add explicit
concurrent-futures 1.3.0 to resolve dependency constraint conflicts.

Upgrade networking and DI stack: Retrofit 3.0.0, OkHttp 5.4.0, and
Koin 4.2.2. Move OkHttp timeouts to kotlin.time.Duration, add
HttpClients.timedBuilder() for shared client configuration, and apply
timeouts to the in-app updater GitHub client.

Koin 4: use org.koin.core.module.dsl.viewModel and remove deprecated
KoinAndroidContext wrapper. Remove obsolete KoinExperimentalAPI
compiler opt-in.

Fix in-app updater install flow (ACTION_VIEW + FileProvider, activity
context handling), version normalization, and UseKtx toUri() usage.
Adapt to OkHttp 5 non-null Response.body.

Add monochrome launcher icon layer for themed icons on Android 13+.
Clean up adaptive icon resources and remove obsolete mipmap/WebP
assets and manifest tools attributes.

Harden coroutine error handling: rethrow CancellationException in
repositories, view models, and WeatherWidgetWorker instead of
swallowing it in broad catch blocks.

Compose/lint cleanups: derivedStateOf, lambda offset, mutableFloatStateOf,
modifier ordering, Moshi @param:Json, Kotlin 2.4 redundant conversions,
and DominicaWeatherParserTest safe-call warnings.

Expand ProGuard rules for Koin 4 core and Retrofit API interfaces.
Scope time4j R8 keep-rule ignore to the library artifact.

Earlier cleanup: remove dead code/deps/empty folders, trim unused
Compose modules, and fix obsolete SDK branches (minSdk 31).

App version: 1.6.2 (versionCode 8).

Verified: assembleDebug, assembleRelease, testDebugUnitTest, and
lintDebug all pass with zero lint warnings.

+++ app/src/main/java/com/weatherpossum/app/presentation/components/ExpressiveIndicators.kt
@@ -199,7 +199,7 @@ fun WavyCircleProgressIndicator(
                 
                 while (angle <= startAngle + totalAngle) {
                     val angleRad = Math.toRadians(angle.toDouble())
-                    val waveOffset = amplitudePx * sin((angleRad * waveFrequency + phase).toDouble()).toFloat()
+                    val waveOffset = amplitudePx * sin(angleRad * waveFrequency + phase).toFloat()
                     val currentRadius = baseRadius + waveOffset
                     
                     val x = center.x + currentRadius * cos(angleRad).toFloat()
